Refactoring FragmentReviewAddEdit and Add and Edit dialogs to become more flexible through the use of an Input Handler object.
At the moment there is a separate FragmentReview and DialogAddReviewData for each GVType. Trying to change this by splitting out the specific GVType handling logic into InputHandlerReviewData objects that know how to deal with their own GVType so that FragmentReview and Add/Edit dialogs can be abstracted into just one class. Aim is that each of these then just fowards user input to the appropriate handler.

Working for GVTag.