Changed ReviewReview framework to work with attach android pattern
ReviewViews are packed by the Administrator into the intent starting the screen. FragmentViewReview passes the intent that started its parent activity back to the Administrator which then passes the ReviewView it needs to show. At this point fragment sets itself ("attaches") on the ReviewView which in turn attaches itself to its actions. Actions then have access to activity etc.